Move CSSSelectorParserContext to its own file https://bugs.webkit.org/show_bug.cgi?id=265220 rdar://118700294
Reviewed by Jean-Yves Avenard.

Allow constructing CSSSelectorParserContext outside of CSSSelectorParser without including the entirety of CSSSelectorParser.h

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/parser/CSSSelectorParser.cpp:
(WebCore::CSSSelectorParserContext::CSSSelectorParserContext): Deleted.
(WebCore::add): Deleted.
* Source/WebCore/css/parser/CSSSelectorParser.h:
(WebCore::CSSSelectorParserContextHash::hash): Deleted.
(WebCore::CSSSelectorParserContextHash::equal): Deleted.
(WTF::HashTraits<WebCore::CSSSelectorParserContext>::constructDeletedValue): Deleted.
(WTF::HashTraits<WebCore::CSSSelectorParserContext>::isDeletedValue): Deleted.
(WTF::HashTraits<WebCore::CSSSelectorParserContext>::emptyValue): Deleted.
* Source/WebCore/css/parser/CSSSelectorParserContext.cpp: Added.
(WebCore::CSSSelectorParserContext::CSSSelectorParserContext):
(WebCore::add):
* Source/WebCore/css/parser/CSSSelectorParserContext.h: Added.
(WebCore::CSSSelectorParserContextHash::hash):
(WebCore::CSSSelectorParserContextHash::equal):
(WTF::HashTraits<WebCore::CSSSelectorParserContext>::constructDeletedValue):
(WTF::HashTraits<WebCore::CSSSelectorParserContext>::isDeletedValue):
(WTF::HashTraits<WebCore::CSSSelectorParserContext>::emptyValue):

Canonical link: https://commits.webkit.org/271035@main
