+++ Source/WebCore/Headers.cmake
@@ -875,6 +875,7 @@ set(WebCore_PRIVATE_FRAMEWORK_HEADERS
     css/parser/CSSParserToken.h
     css/parser/CSSParserTokenRange.h
     css/parser/CSSSelectorParser.h
+    css/parser/CSSSelectorParserContext.h
 
     css/query/GenericMediaQueryTypes.h
     css/query/MediaQuery.h
